+++ assembly/src/tests.rs
@@ -3313,7 +3313,7 @@ prop_compose! {
 
         let manifest = PackageManifest::new(exports).with_dependencies(manifest.dependencies().cloned());
 
-        Package { name, mast, manifest, account_component_metadata_bytes: None }
+        Package { name, version: None, description: None, mast, manifest, sections: Default::default() }
     }
 }