+++ core/src/program.rs
@@ -2,6 +2,8 @@ use alloc::{sync::Arc, vec::Vec};
 use core::fmt;
 
 use miden_crypto::{Felt, WORD_SIZE, Word};
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
 use winter_math::FieldElement;
 use winter_utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable};
 
@@ -21,6 +23,7 @@ use crate::{
 /// execution begins, and a definition of the kernel against which the program must be executed
 /// (the kernel can be an empty kernel).
 #[derive(Clone, Debug, PartialEq, Eq)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct Program {
     mast_forest: Arc<MastForest>,
     /// The "entrypoint" is the node where execution of the program begins.